Change field types in AIInsightPayload to float
Updated data types for various fields in AIInsightPayload to accept float values instead of int.

+++ Code/backend/app/models/insights.py
@@ -6,10 +6,10 @@ class AIInsightPayload(BaseModel):
     source_id: str = Field(...)
 
     vehicle_id: str
-    timestamp_ms: int
+    timestamp_ms: float   # accepts int or float
 
-    fog_decision_critical_class: int
-    fog_decision_actuation_triggered: int
+    fog_decision_critical_class: float
+    fog_decision_actuation_triggered: float
     fog_decision_confidence: float
 
     thermal_brake_margin: float
@@ -17,7 +17,7 @@ class AIInsightPayload(BaseModel):
     thermal_stress_index: float
 
     mechanical_vibration_anomaly_score: float
-    mechanical_dominant_fault_band_hz: int
+    mechanical_dominant_fault_band_hz: float
     mechanical_vibration_rms: float
 
     electrical_charging_efficiency_score: float
@@ -26,9 +26,9 @@ class AIInsightPayload(BaseModel):
     usage_driver_aggression_score: float
     usage_stress_amplification_factor: float
 
-    engine_rul_pct: int
-    brake_rul_pct: int
-    battery_rul_pct: int
+    engine_rul_pct: float
+    brake_rul_pct: float
+    battery_rul_pct: float
 
     fault_primary: str
     fault_contributing_factor: List[str]
@@ -38,7 +38,7 @@ class AIInsightPayload(BaseModel):
 
     recommendation_service_priority: str
     recommendation_suggested_action: str
-    recommendation_safe_operating_limit_km: int
+    recommendation_safe_operating_limit_km: float
 
     trigger_measured_brake_temp_c: float
     trigger_brake_temp_rise_rate: float
